Trip distribution with twitter counts, added a OpenAttractorDataReader that reads numberOftweets or TweetDensity. Also added a new input file for twitter coefficients (non-calibrated)

+++ src/main/java/de/tum/bgu/msm/io/input/readers/PersonsReader.java
@@ -88,7 +88,7 @@ public class PersonsReader extends AbstractCsvReader {
 
 
         //mito uses monthly income, while SILO uses annual income
-        int monthlyIncome_EUR = Integer.parseInt(record[posIncome])/12;
+        double monthlyIncome_EUR = Double.parseDouble(record[posIncome])/12;
         hh.addIncome(monthlyIncome_EUR);
 
         MitoOccupation occupation = null;
